update
You can output some details.
Therefore I changed some programs.

+++ REA2310/PrintForm.cs
@@ -17,24 +17,25 @@ namespace REA2310
         public PrintForm(MainFormModel formData, IData appData)
         {
             InitializeComponent();
-            //this.date = date;
             this.formData = formData;
 
             this.appData = appData;
+
+            // SectionReport.csで必要なデータの参照先を設定
             filePath = this.appData.GetRootDirectoryPath() + @"\REA2310.csv";
 
-            if (File.Exists(filePath))
-            {
-                File.Delete(filePath);
-            }
+            // 重複するため、存在していた場合は削除
+            if (File.Exists(filePath)) File.Delete(filePath);
         }
 
         private void viewer1_Load(object sender, EventArgs e)
         {
+            // コンストラクタで設定したMainFormのデータとfileのパスをController.csへ渡す
             var control = new Controller(formData, filePath);
             control.CreateData();
 
-            SectionReport sectionReport = new SectionReport(control.GetDate(), control.GetBank());
+            // control.CreateDate()で生成したデータをSectionReport.csへ渡す
+            SectionReport sectionReport = new SectionReport(control.GetDate(), control.GetBank(), formData);
             viewer1.LoadDocument(sectionReport);
         }
     }
